+++ common/src/main/java/org/geogebra/common/euclidian/EuclidianController.java
@@ -4253,7 +4253,7 @@ public abstract class EuclidianController {
 	 *
 	 * @return percentage for which we capture point to grid
 	 */
-	final public float getPointCapturingPercentage() {
+	final public double getPointCapturingPercentage() {
 		return getCompanion().getPointCapturingPercentage();
 	}
