+++ common/src/main/java/org/geogebra/common/euclidian/draw/DrawConic.java
@@ -1336,10 +1336,9 @@ public class DrawConic extends Drawable implements Previewable {
 	 */
 	protected void updateParabolaPath() {
 		parabola.reset();
-		parabola.moveTo((float) parpoints[0], (float) parpoints[1]);
-		parabola.curveTo((float) parpoints[2], (float) parpoints[3],
-				(float) parpoints[4], (float) parpoints[5],
-				(float) parpoints[6], (float) parpoints[7]);
+		parabola.moveTo(parpoints[0], parpoints[1]);
+		parabola.curveTo(parpoints[2], parpoints[3], parpoints[4], parpoints[5],
+				parpoints[6], parpoints[7]);
 	}
 
 	@Override
